+++ apps/calendly/src/connectors/calendly/auth.test.ts
@@ -1,10 +1,8 @@
- 
-
 import { http } from 'msw';
 import { describe, expect, test, beforeEach } from 'vitest';
 import { server } from '@elba-security/test-utils';
 import { env } from '@/common/env';
-import { CalendlyError } from '../common/error';
+import { CalendlyError } from './common/error';
 import { getToken, getRefreshToken } from './auth';
 
 const validCode = '1234';
