+++ apps/calendly/src/connectors/calendly/users.test.ts
@@ -2,7 +2,7 @@ import { http } from 'msw';
 import { describe, expect, test, beforeEach } from 'vitest';
 import { server } from '@elba-security/test-utils';
 import { env } from '@/common/env';
-import { CalendlyError } from '../common/error';
+import { CalendlyError } from './common/error';
 import type { CalendlyUser } from './users';
 import { getUsers, deleteUser } from './users';
