+++ apps/calendly/src/inngest/functions/token/refresh-token.test.ts
@@ -6,7 +6,7 @@ import { db } from '@/database/client';
 import { organisationsTable } from '@/database/schema';
 import * as authConnector from '@/connectors/calendly/auth';
 import { encrypt, decrypt } from '@/common/crypto';
-import { CalendlyError } from '@/connectors/common/error';
+import { CalendlyError } from '@/connectors/calendly/common/error';
 import { refreshToken } from './refresh-token';
 
 const newTokens = {
