+++ apps/calendly/src/inngest/middlewares/unauthorized-middleware.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, test, vi } from 'vitest';
 import { NonRetriableError } from 'inngest';
 import { db } from '@/database/client';
 import { organisationsTable } from '@/database/schema';
-import { CalendlyError } from '@/connectors/common/error';
+import { CalendlyError } from '@/connectors/calendly/common/error';
 import { unauthorizedMiddleware } from './unauthorized-middleware';
 
 const organisation = {
